Merging logo onto header background...
Originally this was a PNG that floated above the background
with transparency.  Unfortunately this is not properly
supported by older versions of IE, and looked a mess.

This change will now stop people from getting the logo by
right-clicking and saving.  I don't know if this is a good
or bad thing.  The logo is still available for download if
explicitly requested.